backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
httpx==0.27.2
python-dotenv==1.0.1
pdfplumber==0.11.4
gTTS==2.5.3
pydantic==2.9.2
python-multipart==0.0.9
yt-dlp>=2025.0.0
ffmpeg-python==0.2.0
imageio-ffmpeg==0.5.1
youtube-transcript-api
faster-whisper
fpdf2
python-docx
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
